+++ src/views/feed/yandex.php
@@ -16,11 +16,11 @@
         <title><?=htmlspecialchars($item['name'],ENT_QUOTES,\Yii::$app->charset);?></title>
         <link><?=$item['url'];?></link>
         <?php if ($item['category']):?>
-            <category><![CDATA['<?=htmlspecialchars($item['category'],ENT_QUOTES,\Yii::$app->charset);?>']]></category>
+            <category><![CDATA[<?=htmlspecialchars($item['category'],ENT_QUOTES,\Yii::$app->charset);?>]]></category>
         <?php endif;?>
         <description><![CDATA[<?=$item['text'];?>]]></description>
         <?php if (is_file(ROOT_DIR.'/frontend/web'.$item['img_src'])):?>
-        <enclosure url="<?=$item['img_src'];?>" type="<?= \yii\helpers\FileHelper::getMimeType(ROOT_DIR.'/frontend/web'.$item['img_src']);?>"/>
+        <enclosure url="<?=\frontend\helpers\Url::to($item['img_src'], 'https');?>" type="<?= \yii\helpers\FileHelper::getMimeType(ROOT_DIR.'/frontend/web'.$item['img_src']);?>"/>
         <?php endif;?>
         <pubDate><?= $item['dt_start'];?></pubDate>
         <yandex:full-text><?= $item['full-text'];?></yandex:full-text>