+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 CHANGELOG
 ==============
 
+1.0.6
+-----------------
+  * Add Ukrnet feed with filter related property
+
+1.0.5
+-----------------
+  * Add date rss converting with time zone settings
+
+1.0.4
+-----------------
+  * Add Yandex feed
+
+1.0.3
+-----------------
+  * Add feed title description image from tree
+
+1.0.2
+-----------------
+  * Add text-full feed
+
 1.0.0-alpha
 -----------------
   * Can be used